feat: unit tests and beginning int. tests (#15)
Unit tests and beginning integration tests.

There ware two refactors needed for this. Azure's Go SDK uses a paging approach where mocking it is tricky. So I made a facade client that is easier to mock and refactored the main code to use it too. Also, there was code in the validator that got a role lookup map from Azure. I had to change this to a provider as another dependency. With these changes, the unit tests could work without talking to Azure.