docs: principal ID and rules in readme (#222)
## Issue
There wasn't much guidance on how to specify the correct `principalId`
in the spec for RBAC rule. Azure is complex when it comes to this. There
are four GUIDs users see in the Azure Portal, where three of them are
unique to each Entra ID identify they create. We need to guide them.

## Description
Clarifies how principalID works in the docs. This includes the docstring
generated by kubebuilder and some additions to the readme that go into
detail about it, including screenshots.

Because there is now a readme section on each rule in detail, there is a
part where it describes the community gallery image rule too. But there
isn't much to say about that rule right now.

Also removes some unneeded `MaxLength` annotations (because they aren't
used in other kubebuilder validations where long strings are
problematic).

Note that unlike in the PR
https://github.com/validator-labs/validator-plugin-azure/pull/219, the
`ActionStr` type and its `MaxLength` annotation must remain. This is
because the RBAC rule forbids wildcards in specified permissions and
kubebuilder validation here validates this, which is a type of
kubebuilder validation where long strings are problematic. We need
`ActionStr` because we need to place the `MaxLength` annotation on the
list of actions in a way where it is applied to each string in a slice
of strings, and this is the recommended way to do that right now.

---------

Signed-off-by: Matt Welke <matt.welke@spectrocloud.com>