ci: add workflow for duplicating issues to validatorctl (#259)
## Issue

## Description
Adding a workflow to duplicate issues from the plugin repo into
validatorctl.

When a new issue is created which requires work in both the plugin and
validatorctl repos, it should be labeled with `validatorctl`. A copy
will be automatically made in the `validatorctl` repo, assigned to the
same person, and added to the `validator` project board.