feat: quota rule (#271)
## Issue
Resolves #183 

## Description
Adds support for the rule. Here's what it looks like:

```yaml
apiVersion: validation.spectrocloud.labs/v1alpha1
kind: AzureValidator
metadata:
  name: azurevalidator-quota-one-resource-set-one-resource
spec:
  auth:
    implicit: false
    secretName: azure-creds
  quotaRules:
  - name: rule-1
    resourceSets:
    - scope: /subscriptions/5f6df17d-dc8f-45e0-ba9f-0d5601c70df8/providers/Microsoft.Compute/locations/westus
      resources:
      - name: availabilitySets # defaults to a quota limit of 2500 in new Azure subscriptions
        buffer: 1000
```

### Details

Azure's terminology:

1. Quota limit: The limit for a resource.
2. Usage: The current amount of that resource provisioned.

Users specify the resources they want to check quota limits and usages
for. For each resource, they can also set a buffer. The system checks if
the current usage plus the buffer exceeds the limit. If it doesn't, the
validation passes. If it does, the validation fails. Buffer is required.
If users don't care about a buffer, they can set it to 0.

In Azure, quota limits and usages are associated with specific resource
names and these resources are organized within scopes. Therefore, when
using this rule, users need to define the scope they are interested in
and then specify the resources within that scope whose limits and usages
they want to monitor. If there's a mismatch because the user specified a
resource under a scope that doesn't belong to that scope, it will result
in a validation failure with this mismatch reported in a failure
message.

Here's how we use the API. Scope is a parameter in Azure's [Quota
Service API](https://learn.microsoft.com/en-us/rest/api/quota/), which
is what powers this. For each resource set the user specifies, we make a
"list quota limits for scope" API call and a "list usages for scope" API
call. We then iterate through the resources the user specifies in the
resource set (the resources for that scope) and check the limits and
usages for that resource.

This ensures there are only two API calls made per resource set, not two
API calls made per resource. Validation should complete quickly as long
as users group their resources within scopes as much as possible,
defining as few resource sets as possible.

A nice thing about how Azure's API works is that it uses a single
endpoint for all quota types and a single endpoint for all usage types
(it's just the scope that changes). So in the future, if they add a new
quota type, this API can continue to be used as is (just with the new
scope). That means once we merge this, we won't need to update it to add
support for new quota types Azure introduces.

A not so nice thing about how Azure's API works is that they don't seem
to list all the resources for each resource provider. The only way I
could think of to find this list was to make my own API call to my Azure
subscription to see which quota limits I happened to have at that
moment. Then, I treat what comes back from this API call as the list,
until it becomes out of date. I've left instructions in the README for
users to do exactly this when they want to find a list of resources they
can specify in rules.

---------

Signed-off-by: Matt Welke <matt.welke@spectrocloud.com>